fix(ru-client, ru-client-model, ru-service-model): migrate all id to Long, all data from string to LocalDate

+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/AddressModel.kt
@@ -1,10 +1,10 @@
 package com.fs.client.ru
 
 data class AddressModel(
-    val id: Int,
-    val cityId: Int?,
-    val apartment: Int?,
-    val building: Int?,
-    val house: Int?,
+    val id: Long,
+    val cityId: Long?,
+    val apartment: Long?,
+    val building: Long?,
+    val house: Long?,
     val street: String?
 )
