+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/CityModel.kt
@@ -1,7 +1,7 @@
 package com.fs.client.ru
 
 data class CityModel(
-    val id: Int,
-    val countryId: Int,
+    val id: Long,
+    val countryId: Long,
     val name: String
 )
