+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/CountryModel.kt
@@ -4,7 +4,7 @@ import com.fs.client.ru.enums.CountryNameModel
 import com.fs.client.ru.enums.CurrencyModel
 
 data class CountryModel(
-    val code: Int,
+    val code: Long,
     val currency: CurrencyModel,
     val name: CountryNameModel
 )
