+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/OfficeModel.kt
@@ -1,8 +1,8 @@
 package com.fs.client.ru
 
 data class OfficeModel(
-    val id: Int?,
-    var addressId: Int?,
-    val companyId: Int?,
+    val id: Long?,
+    var addressId: Long?,
+    val companyId: Long?,
     val phoneNumber: String?
 )
