+++ fs-platform-ru/backend/domain/ru-client-model/src/main/kotlin/com/fs/client/ru/PartnerModel.kt
@@ -1,5 +1,5 @@
 package com.fs.client.ru
 
 data class PartnerModel(
-    val partnerId: Int
+    val partnerId: Long
 )
