+++ fs-platform-ru/backend/domain/ru-service-model/src/main/kotlin/com/fs/service/ru/ServiceModel.kt
@@ -1,7 +1,7 @@
 package com.fs.service.ru
 
 data class ServiceModel(
-    val id: Int?,
-    val price: Int?,
+    val id: Long?,
+    val price: Long?,
     val name: String?
 )
